requirements.txt
dora-search
hydra-core
transformers
sentence-transformers
datasets
spacy
tabulate
pandas
pyyaml
tqdm
prettytable
